Refuse to consult strategy or issue directives if no data
Refactors code from SimpleProportionalStrategy into the
SchedulerDirector class, since if data is missing the scheduler should
treat that as an error condition and refuse to invoke the strategy or
issue directives.

It feels more appropriate for this code to live in the scheduler rather
than a specific strategy.

Includes a unit test and unit/functional test refactoring to accomodate
changes.